% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDER_GE.R
\name{runLDER_GE}
\alias{runLDER_GE}
\title{Main function}
\usage{
runLDER_GE(
  assoc,
  n.gwis,
  path,
  LD.insample = T,
  n.ld,
  method = "lder",
  cores = 10,
  type = "jack",
  size_num = 200
)
}
\arguments{
\item{assoc}{GWIS summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{n.gwis}{The sample size of the GWIS summary statistics}

\item{path}{The path of LD panel directory}

\item{LD.insample}{T/F, whether the LD reference is estimated with the GWIS cohort (T) or external reference panel (e.g. 1000 Genome Project) (F)}

\item{n.ld}{The sample size of the LD reference}

\item{method}{'lder', 'ldsc', or 'both'}

\item{cores}{The number of cores for computation in parallel}

\item{type}{if 'jack' then conducts delete-wise block jackknife. if 'none' then does not do inference}

\item{size_num}{Number of blocks for jackknife}
}
\description{
Run LDER-GE
}
